warmup -> train.warmup, DRY
