+++ configs/dstart.yaml
@@ -5,6 +5,7 @@ train:
   batch_size: 16
   distributed: False
   label_smoothing: 0.1
+  warmup: 4000
 
 dataset_name: 'roszcz/maestro-v1-sustain'
 target: dstart
@@ -25,7 +26,6 @@ dataset:
     velocity: 3
 
 device: "cuda:0"
-warmup: 3000
 log_frequency: 10
 run_name: midi-transformer-${now:%Y-%m-%d-%H-%M}
 project: "midi-transformer"
